Rename eighty folder to m80v1
