+++ keyboards/mode/m80v1/m80s/info.json
